Covert hover icons to named exports
