+++ client-v2/src/shared/images/icons-hover/index.js
@@ -1,17 +1,6 @@
-import View from './icon-hover-view.svg';
-import Ophan from './icon-hover-ophan.svg';
-import Copy from './icon-hover-copy.svg';
-import Paste from './icon-hover-paste-sublink.svg';
-import Clipboard from './icon-hover-clipboard.svg';
-import Delete from './icon-hover-delete.svg';
-
-const hoverActionIcons = {
-  view: View,
-  ophan: Ophan,
-  copy: Copy,
-  paste: Paste,
-  clipboard: Clipboard,
-  delete: Delete
-};
-
-export default hoverActionIcons;
+export { default as View } from './icon-hover-view.svg';
+export { default as Ophan } from './icon-hover-ophan.svg';
+export { default as Copy } from './icon-hover-copy.svg';
+export { default as Paste } from './icon-hover-paste-sublink.svg';
+export { default as Clipboard } from './icon-hover-clipboard.svg';
+export { default as Delete } from './icon-hover-delete.svg';
